app/requirements.txt
aiortc==1.2.0
av==8.0.3
numpy==1.20.1
pandas
Pillow
playsound
torch
torchvision
scikit-image
opencv_python==4.5.1.48
streamlit==0.78.0
streamlit_webrtc==0.7.2
typing_extensions==3.7.4.3
base64
streamlit_player